Latest Changes

+++ ClientApp/src/app/Admin/UserRole/partial-user_role/partial-user_role.component.ts
@@ -3,17 +3,16 @@ import { ApiResponseModel } from 'src/Helper/api-response-model';
 import { user_role } from 'src/Model/user_role';
 import { CommonHelper } from 'src/Helper/CommonHelper';
 import { FormGroup, FormBuilder, FormControl, Validators } from '@angular/forms';
-import { DynamicDialogRef, DynamicDialogConfig,SelectItem } from 'primeng/api';
+import { DynamicDialogRef, DynamicDialogConfig, SelectItem, ConfirmationService } from 'primeng/api';
 import { CommonService } from 'src/Service/Common.service';
-
 
 @Component({
   selector: 'app-partial-user_role',
   templateUrl: './partial-user_role.component.html',
-  styleUrls: ['./partial-user_role.component.scss']
+  styleUrls: ['./partial-user_role.component.scss'],
+  providers: [ConfirmationService]
 })
 export class PartialUserRoleComponent implements OnInit {
-
   UserRoleData: user_role;
   UserRoleForm: FormGroup;
 
@@ -22,20 +21,23 @@ export class PartialUserRoleComponent implements OnInit {
     public ref: DynamicDialogRef,
     public config: DynamicDialogConfig,
     public commonservice: CommonService,
-    private formbuilder: FormBuilder
+    private formbuilder: FormBuilder,
+    private confirmationService: ConfirmationService
   ) {
     this.UserRoleData = this.config.data;
   }
+
   ngOnInit() {
     this.UserRoleForm = this.formbuilder.group({
       name: new FormControl('', Validators.compose([
         Validators.required
       ])),
-       landing_page: new FormControl('', Validators.compose([
+      landing_page: new FormControl('', Validators.compose([
         Validators.nullValidator
       ])),
     });
   }
+
   UserRoleValidationMessages = {
     'name': [
       { type: 'required', message: 'Landing Page is required.' },
@@ -65,26 +67,37 @@ export class PartialUserRoleComponent implements OnInit {
       }
     );
   }
-  Delete() {
-    this.helper.ShowSpinner();
-    this.commonservice.Delete(this.UserRoleData.id, "UserRoleDelete").subscribe(
-      (res: ApiResponseModel) => {
-        if (res.Type == "S") {
-          this.helper.SucessToastr(res.Message, "User Role");
-          this.ref.close(true);
-        } else {
-          this.helper.ErrorToastr(res.Message, "User Role");
-        }
-      },
-      error => {
-        this.helper.HideSpinner();
-        this.helper.ErrorToastr(error, "Error");
-      },
-      () => {
-        this.helper.HideSpinner();
-      }
-    );
-  }
 
+  Delete() {
+    debugger
+    this.confirmationService.confirm({
+      message: 'Do you want to delete this '+ this.UserRoleData.name + ' record?',
+      header: 'Delete Confirmation',
+      icon: 'pi pi-info-circle',
+      accept: () => {
+        this.helper.ShowSpinner();
+        this.commonservice.Delete(this.UserRoleData.id, "UserRoleDelete").subscribe(
+          (res: ApiResponseModel) => {
+            if (res.Type == "S") {
+              this.helper.SucessToastr(res.Message, "User Role");
+              this.ref.close(true);
+            } else {
+              this.helper.ErrorToastr(res.Message, "User Role");
+            }
+          },
+          error => {
+            this.helper.HideSpinner();
+            this.helper.ErrorToastr(error, "Error");
+          },
+          () => {
+            this.helper.HideSpinner();
+          }
+        );
+      },
+      reject: () => {
+        //this.msgs = [{ severity: 'info', summary: 'Rejected', detail: 'You have rejected' }];
+      }
+    });
+  }
 
 }
